% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi.R
\name{cover}
\alias{cover}
\title{produces cover and soil surface by symbol (species or other) by plot or line}
\usage{
cover(x, type = "general", by = "line")
}
\arguments{
\item{x}{data type list produced by \code{load_data()}.}

\item{type}{can be \code{"general"} for generalized information by line or \code{"detailed"} for symbol level data.}

\item{by}{if type is \code{"detailed"} this can be either \code{"plot"} or \code{"line"}. Default is \code{"line"}}
}
\description{
produces cover and soil surface by symbol (species or other) by plot or line
}
