% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi.R
\name{dima_cover}
\alias{dima_cover}
\title{produces cover by type by plot}
\usage{
dima_cover(data, by = "line")
}
\arguments{
\item{data}{type list produced by \code{read_dima()}.}
}
\description{
produces cover by type by plot
}
