% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap.R
\name{dima_gap}
\alias{dima_gap}
\title{returns total gap and gap percent by plot.}
\usage{
dima_gap(x)
}
\arguments{
\item{x}{data of type list created by \code{read_dima())}}
}
\description{
returns total gap and gap percent by plot.
}
