% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi.R
\name{dima_heights}
\alias{dima_heights}
\title{produces herbaceous and woody heights by type (woody and herbaceous as well as big sage) by plot.}
\usage{
dima_heights(data, raw = FALSE)
}
\arguments{
\item{data}{type list produced by \code{read_dima()}.}

\item{raw}{if \code{TRUE} returns raw dataset not summarized.}
}
\description{
produces herbaceous and woody heights by type (woody and herbaceous as well as big sage) by plot.
}
