% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap.R
\name{gap}
\alias{gap}
\title{returns total gap and gap percent by plot of survey123 data.}
\usage{
gap(x)
}
\arguments{
\item{x}{data of type list created by \code{load_data()}}
}
\description{
returns total gap and gap percent by plot of survey123 data.
}
