% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi.R
\name{height}
\alias{height}
\title{produces heights for all plots by line or plot.}
\usage{
height(x, by = "line")
}
\arguments{
\item{x}{data type list produced by \code{load_data()}.}

\item{by}{type data will be returned as. Can return by \code{"line"} (default) or \code{"plot"}}
}
\description{
produces heights for all plots by line or plot.
}
