% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_qc.R
\name{qa_qc}
\alias{qa_qc}
\title{Runs all qc functions and returns a list of tibbles, each with qc checks. The tables returned are \code(unknown_symbol_check) and \code(sp_rich_check) for now.}
\usage{
qa_qc(x)
}
\arguments{
\item{x}{type list produced by \code{read_dima()} function.}
}
\description{
Runs all qc functions and returns a list of tibbles, each with qc checks. The tables returned are \code(unknown_symbol_check) and \code(sp_rich_check) for now.
}
