% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sage_calcs.R
\name{sage_cover}
\alias{sage_cover}
\title{calculates sage cover percent.}
\usage{
sage_cover(x, by = "line")
}
\arguments{
\item{x}{type list produced by \code{read_dima()}.}

\item{by}{if type is \code{"detailed"} this can be either \code{"plot"} or \code{"line"}. Default is \code{"line"}}
}
\value{
a tibble of sage cover percentages.
}
\description{
calculates sage cover percent.
}
\examples{
library(AIMtools)
# Make sure you are connected to AGOL Online
library(arcgisbinding)
arc.check_product()

# Load data from AGOL online
agol_data<-load_data()

sage_cover(agol_data, by="line")
}
