% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_richness.R
\name{sp_rich}
\alias{sp_rich}
\title{Analyze species richness data from AIM data hosted on ArcGIS online.}
\usage{
sp_rich(data, func = "count", unknowns = TRUE)
}
\arguments{
\item{data}{a list of data frames created by \code{load_data()}.}

\item{func}{analysis function. Can be character or vector of characters. Options are "count".}

\item{unknowns}{\code{TRUE} or \code{FALSE}.  If \code{TRUE} identified unknowns in database will be merged into count data.}
}
\description{
Analyze species richness data from AIM data hosted on ArcGIS online.
}
