% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_richness.R
\name{sp_rich_dima}
\alias{sp_rich_dima}
\title{Analyze species richness from AIM data in a DIMA database.
#'}
\usage{
sp_rich_dima(x, func = "count")
}
\arguments{
\item{x}{a list of data frames created by \code{read_dima()}.}

\item{func}{analysis function. Can be character or vector of characters. Options are "count".}
}
\description{
Analyze species richness from AIM data in a DIMA database.
#'
}
