% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_qc.R
\name{sp_rich_qc}
\alias{sp_rich_qc}
\title{produces a table with a variety of species richness related qc checks, with information on what the problem is and how to fix it.}
\usage{
sp_rich_qc(x, output = FALSE)
}
\arguments{
\item{x}{type list produced by \code{read_dima()} function.}
}
\description{
produces a table with a variety of species richness related qc checks, with information on what the problem is and how to fix it.
}
