% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknowns.R
\name{unknown_count}
\alias{unknown_count}
\title{counts number of unknowns that have been identified.}
\usage{
unknown_count(x, by = FALSE)
}
\arguments{
\item{x}{type list produced by \code{read_dima()}.}

\item{by}{default is FALSE.  Can count by unknown type if by is set to \code{type}}
}
\value{
A table of the number of unkonwns with a FinalCode identified.
}
\description{
counts number of unknowns that have been identified.
}
\examples{
library(AIMtools)
# Make sure you are connected to AGOL Online
library(arcgisbinding)
arc.check_product()

# Load data from AGOL online
agol_data<-load_data()

#Table of identified unknowns.
unknown_count(agol_data)

#Table of unknowns by type ("AF", "PF", "SH", etc.).
unknown_count(agol_data, by="type")

}
